refactor: centralize list_vaults helper
