+++ src/tino_storm/ingest/search.py
@@ -7,6 +7,7 @@ from typing import Any, Iterable, List, Dict, Optional
 
 import chromadb
 
+from .utils import list_vaults  # noqa: F401
 from ..security import (
     get_passphrase,
     encrypt_parquet_enabled,
@@ -17,25 +18,6 @@ from ..security.encrypted_chroma import EncryptedChroma
 from ..retrieval.rrf import reciprocal_rank_fusion
 from ..retrieval.scoring import score_results
 from ..retrieval.bayes import add_posteriors
-
-
-def list_vaults(root: Optional[str] = None) -> list[str]:
-    """Return available vault directories under ``root``.
-
-    If ``root`` is not provided, the ``STORM_VAULT_ROOT`` environment variable is
-    consulted. If that is unset, the default ``~/.tino_storm/research`` directory
-    is used. Only sub-directories are returned and the result is sorted
-    alphabetically.
-    """
-
-    root_path = Path(
-        root or os.environ.get("STORM_VAULT_ROOT") or Path.home() / ".tino_storm" / "research"
-    ).expanduser()
-
-    if not root_path.exists():
-        return []
-
-    return sorted(p.name for p in root_path.iterdir() if p.is_dir())
 
 
 def search_vaults(
